fix(graph): resolve some todos
