+++ src/graph/model.ts
@@ -2,6 +2,7 @@ import { JoinRatesData } from '../joinRates/model'
 import { Dataset as ProfilerData } from '../profiler/model'
 import { QualityData } from '../quality'
 import { Node as Node0, Edge } from 'react-flow-renderer'
+import { RawMetricsDataItemData } from '../metrics/model'
 
 export interface DataPayload {
   label: string
@@ -49,7 +50,7 @@ export interface Table {
   dp_result?: ProfilerData
   dq_result?: QualityData
   dq_join_results?: JoinRatesData[]
-  metrics?: any // TODO: use proper type
+  metrics?: RawMetricsDataItemData[]
   metadata?: {
     [key: string]: string | number
   }
